requirements.txt
# ── Core LangChain 0.3.x (los notebooks usan esta API) ──────────────────────
langchain>=0.3.14,<1.0.0
langchain-community>=0.3.14,<1.0.0
langchain-google-genai>=2.0.8
langchain-experimental>=0.3.4,<1.0.0
langchain-groq>=0.2.3
langgraph>=0.2.60,<1.0.0

# ── Vector store ─────────────────────────────────────────────────────────────
faiss-cpu

# ── Document loaders ─────────────────────────────────────────────────────────
pypdf>=5.1.0
arxiv>=2.1.3

# ── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers>=3.3.1

# ── Knowledge Graph ──────────────────────────────────────────────────────────
rdflib>=7.1.1
SPARQLWrapper>=2.0.0

# ── Web search (fallback) ────────────────────────────────────────────────────
tavily-python>=0.5.0
langchain-tavily

# ── Environment & utilities ───────────────────────────────────────────────────
python-dotenv>=1.0.1
requests>=2.32.3
jupyter
ipykernel

# ── Métricas ─────────────────────────────────────────────────────────────────
numpy
pandas
